@@ -1,9 +1,17 @@
-import { Component, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { FormControl, UntypedFormControl } from '@angular/forms';
 import { Router } from '@angular/router';
-import { Subject, switchMap, takeUntil, tap } from 'rxjs';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
+import { Subject, switchMap, take, takeUntil, tap } from 'rxjs';
 import { CartApiService } from 'src/app/@core/api/carts-api.service';
+import { EventApiService } from 'src/app/@core/api/events-api.service';
+import { TeamApiService } from 'src/app/@core/api/team-api.service';
+import { IDropdownFilters } from 'src/app/@core/models/base.model';
+import { EStatusCart } from 'src/app/@core/models/cart.model';
+import { ITeam } from 'src/app/@core/models/team.model';
 import { AuthService } from 'src/app/@core/services/auth.service';
+import { BudgetModalComponent } from './budget-modal/budget-modal.component';
 
 @Component({
   selector: 'app-personal-cart-list',
@@ -11,42 +19,84 @@ import { AuthService } from 'src/app/@core/services/auth.service';
   styleUrls: ['./personal-cart-list.component.scss'],
 })
 export class PersonalCartListComponent implements OnInit {
-  carts: any = [];
+  events: any = [];
+  teams: Partial<ITeam>[] = [];
   totalRecords: number = 0;
   page: number = 0;
   size: number = 10;
   filter: string = '';
+  startDateFilter?: Date;
+  endDateFilter?: Date;
   sort: any = null;
   loading: boolean = false;
   currentUser: any;
+  EStatusCart = EStatusCart;
 
-  teamsFilter: FormControl = new FormControl([]);
+  teamHomeFilter = new UntypedFormControl('');
+  dateFilter = new UntypedFormControl([new Date(), null]);
+  teamsFilter: FormControl = new FormControl('');
 
-  carts$: Subject<void> = new Subject();
+  events$: Subject<void> = new Subject();
+  teams$: Subject<string> = new Subject();
   unsubscribe$: Subject<void> = new Subject();
+
+  ref!: DynamicDialogRef;
+
+  get teamHomeFilterId() {
+    return this.teamHomeFilter.value ? this.teamHomeFilter.value.id : null;
+  }
 
   constructor(
     private cartsApiService: CartApiService,
+    private eventsApiService: EventApiService,
+    private teamApiService: TeamApiService,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService,
     private router: Router,
-    private authService: AuthService
+    private authService: AuthService,
+    private cd: ChangeDetectorRef,
+    public dialogService: DialogService
   ) {
-    this.carts$
+    this.events$
       .pipe(
         takeUntil(this.unsubscribe$),
         switchMap(() =>
-          this.cartsApiService.findAll({
+          this.eventsApiService.findAll({
             page: this.page + 1,
             take: this.size,
-            ...(this.filter ? { name: this.filter } : {}),
+            ...(this.filter ? { away: this.filter } : {}),
+            ...(this.teamHomeFilterId ? { home: this.teamHomeFilterId } : {}),
+            ...(this.startDateFilter
+              ? { startDate: this.startDateFilter.toISOString() }
+              : {}),
+            ...(this.endDateFilter
+              ? { endDate: this.endDateFilter.toISOString() }
+              : {}),
             ...(this.sort && this.sort.field
               ? { sortField: this.sort.field, sortOrder: this.sort.order }
               : {}),
           })
         ),
         tap(({ data, total }) => {
-          this.carts = [...data];
+          this.events = [...data];
           this.totalRecords = total;
           this.loading = false;
+        })
+      )
+      .subscribe();
+
+    this.teams$
+      .pipe(
+        takeUntil(this.unsubscribe$),
+        switchMap((name) =>
+          this.teamApiService.findAll({
+            page: 1,
+            take: 50,
+            ...(name ? { name } : {}),
+          })
+        ),
+        tap(({ data }) => {
+          this.teams = [...data];
         })
       )
       .subscribe();
@@ -54,28 +104,132 @@ export class PersonalCartListComponent implements OnInit {
     this.currentUser = this.authService.currentUser;
   }
 
-  ngOnInit() {
-    this.loadCarts();
-  }
+  ngOnInit() {}
 
   ngOnDestroy(): void {
     this.unsubscribe$.next();
     this.unsubscribe$.complete();
   }
 
-  loadCarts() {
-    this.carts = [];
-    this.loading = true;
-    this.carts$.next();
+  ngAfterViewInit() {
+    this.cd.detectChanges();
   }
 
-  openCart(cart: any) {
+  loadCarts() {
+    this.events = [];
+    this.loading = true;
+    this.events$.next();
+  }
+
+  openViewCart(cart: any) {
     this.router.navigate(['/personal/carts', cart.id]);
   }
 
-  onChangePage(event: any) {}
+  openCreateCart(event: any) {
+    this.router.navigate(['/personal/carts/new', event.id]);
+  }
+
+  onChangePage(event: any) {
+    this.page = event.first! / event.rows! || 0;
+
+    if (event.sortField) {
+      this.sort = {
+        field: event.sortField,
+        order: event.sortOrder,
+      };
+    } else {
+      this.sort = null;
+    }
+
+    this.loadCarts();
+  }
 
   onApplyFilters() {
-    console.log('filter', this.teamsFilter.value);
+    // this.filter = this.teamsFilter.value?.id ?? '';
+    // this.startDateFilter =
+    //   this.dateFilter.value && this.dateFilter.value[0]
+    //     ? this.dateFilter.value[0]
+    //     : '';
+    // this.endDateFilter =
+    //   this.dateFilter.value && this.dateFilter.value[1]
+    //     ? this.dateFilter.value[1]
+    //     : '';
+    // this.loadCarts();
+    console.log('this.dateFilter.value', this.dateFilter.value);
+  }
+
+  complete(cart: any) {
+    this.confirmationService.confirm({
+      message:
+        'Una volta confermata la trasferta non potrà più essere modificata e verrà inviata una mail al nostro staff. Si intende procedere?',
+      header: 'Conferma',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.cartsApiService
+          .update(cart.id, { status: EStatusCart.PENDING, onlyStatus: true })
+          .pipe(
+            take(1),
+            tap(() => {
+              this.loadCarts();
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Trasferta confermata',
+                detail: 'Verrà inviata una mail di notifica al nostro staff',
+              });
+            })
+          )
+          .subscribe();
+      },
+    });
+  }
+
+  onAddCart() {
+    console.log('custom cart');
+  }
+
+  onFilterTeam({ filter }: IDropdownFilters) {
+    if (filter) {
+      this.loadFilteredTeams(filter);
+    }
+  }
+
+  loadFilteredTeams(name: string) {
+    this.teams$.next(name);
+  }
+
+  onFilterDate(event: any) {
+    console.log(event);
+  }
+
+  budget(cart: any) {
+    console.log('budget', cart);
+    this.ref = this.dialogService.open(BudgetModalComponent, {
+      header: `Budget`,
+      width: '600px',
+      contentStyle: { overflow: 'visible' },
+      baseZIndex: 10001,
+      data: {
+        budget: cart.budget.budget,
+        isEdit: Boolean(cart.budget.budget),
+      },
+    });
+
+    this.ref.onClose.subscribe((budget: any) => {
+      if (budget) {
+        this.cartsApiService
+          .setBudget(cart.id, budget.budget)
+          .pipe(
+            take(1),
+            tap((data) => {
+              this.loadCarts();
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Budget aggiornato',
+              });
+            })
+          )
+          .subscribe();
+      }
+    });
   }
 }
