@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CartListComponent } from './cart-list/cart-list.component';
 import { PersonalCartCreateComponent } from '../personal/personal-cart/personal-cart-create/personal-cart-create.component';
 import { PersonalCartViewComponent } from '../personal/personal-cart/personal-cart-view/personal-cart-view.component';
+import { CartListComponent } from './cart-list/cart-list.component';
 
 const routes: Routes = [
   {
@@ -14,11 +14,15 @@ const routes: Routes = [
     component: PersonalCartCreateComponent,
     data: {
       isEdit: true,
+      fromBackoffice: true,
     },
   },
   {
     path: ':id',
     component: PersonalCartViewComponent,
+    data: {
+      fromBackoffice: true,
+    },
   },
 ];
 
