@@ -1,15 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import {
-  FormArray,
-  FormControl,
-  FormGroup,
-  UntypedFormGroup,
-} from '@angular/forms';
+import { FormArray, FormControl, FormGroup, Validators } from '@angular/forms';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { Subject, debounceTime, map, switchMap, takeUntil, tap } from 'rxjs';
-import { ActivityApiService } from 'src/app/@core/api/activity-api.service';
-import { EStatusCart } from 'src/app/@core/models/cart.model';
-import { clearFormArray } from 'src/app/@core/utils';
+import { Subject } from 'rxjs';
 import { ModalActivityComponent } from './modal-activity/modal-activity.component';
 
 @Component({
@@ -26,25 +18,17 @@ export class CartCreateActivityComponent {
   loading: boolean = false;
   selectedActivities: any = [];
   filter: string = '';
-  EStatusCart = EStatusCart;
-  _status: EStatusCart = EStatusCart.DRAFT;
 
   ref!: DynamicDialogRef;
 
   searchFilter = new FormControl('');
-  //_activityForm: UntypedFormGroup = new FormGroup({});
 
   @Input() event: any;
   @Input() cityActivities: any = [];
   @Input() activityForm: FormGroup = new FormGroup({});
   @Input() activeIndex: number = 0;
-  @Input() set status(value: EStatusCart) {
-    if (value) {
-      this._status = value;
-      if (value !== EStatusCart.DRAFT) {
-      }
-    }
-  }
+  @Input() maxPax = 0;
+  @Input() isDisabledCart: boolean = false;
 
   @Output() nextStep: EventEmitter<void> = new EventEmitter();
   @Output() prevStep: EventEmitter<void> = new EventEmitter();
@@ -94,6 +78,7 @@ export class CartCreateActivityComponent {
         isEdit: true,
         index: index + 1,
         activitiesList: this.cityActivities,
+        maxPax: this.maxPax,
       },
     });
 
@@ -117,7 +102,7 @@ export class CartCreateActivityComponent {
       startDate: new FormControl(startDate),
       startDateHour: new FormControl(startDate),
       quantity: new FormControl(1),
-      id: new FormControl(null),
+      id: new FormControl(null, Validators.required),
       name: new FormControl(null),
       description: new FormControl(null),
       price: new FormControl(null),
@@ -133,6 +118,7 @@ export class CartCreateActivityComponent {
         activityForm: newActivity,
         index: this.activities.length + 1,
         activitiesList: this.cityActivities,
+        maxPax: this.maxPax,
       },
     });
 
