@@ -4,7 +4,6 @@ import { ConfirmationService, MessageService } from 'primeng/api';
 import { Subject, switchMap, take, takeUntil, tap } from 'rxjs';
 import { CartApiService } from 'src/app/@core/api/carts-api.service';
 import { EStatusCart } from 'src/app/@core/models/cart.model';
-import { compareDates } from 'src/app/@core/utils';
 
 @Component({
   selector: 'app-personal-cart-view',
@@ -22,10 +21,12 @@ export class PersonalCartViewComponent {
   hotel: string = '';
   rooms: any = [];
   roads: any = [];
+  meals: any = [];
   roadsGroupByDate: any[] = [];
   total: number = 0;
   isDownloading: boolean = false;
   EStatusCart = EStatusCart;
+  fromBackoffice = false;
 
   get awayTeam() {
     return this.cart?.event.away;
@@ -59,6 +60,7 @@ export class PersonalCartViewComponent {
     private messageService: MessageService
   ) {
     this.cartId = this.route.snapshot.params['id'];
+    this.fromBackoffice = this.route.snapshot.data['fromBackoffice'];
 
     this.cart$
       .pipe(
@@ -89,35 +91,64 @@ export class PersonalCartViewComponent {
             return group;
           }, []);
 
+          this.meals = this.cart.meals.reduce((group: any, meal: any) => {
+            const { quantity, startDate, description } = meal;
+            const { id: configId, name: configName } = meal.meal;
+
+            const {
+              id: mealId,
+              name: mealName,
+              price: mealPrice,
+            } = meal.meal.meal;
+            const index = group.findIndex((g: any) => g.mealId === mealId);
+            if (index > -1) {
+              group[index].configIds.push({ configId, configName });
+            } else {
+              group.push({
+                mealId,
+                mealName,
+                configIds: [{ configId, configName }],
+                quantity,
+                startDate,
+                description,
+                price: mealPrice,
+              });
+            }
+            return group;
+          }, []);
+
+          this.meals = this.meals.map((m: any) => ({
+            ...m,
+            mealName: `${m.mealName} - ${m.configIds
+              .map((c: any) => c.configName)
+              .join(', ')}`,
+          }));
+
+          this.roads = this.cart?.roads;
+
           const totalAccomodation = this.rooms.reduce((acc: any, room: any) => {
             return acc + room.price * room.quantity * 100;
           }, 0);
           const totalActivity = this.activities.reduce(
             (acc: any, activity: any) => {
-              return acc + activity.activity.price * 100;
+              return acc + activity.quantity * activity.activity.price * 100;
             },
             0
           );
           const totalRoads = this.cart?.roads.reduce((acc: any, road: any) => {
-            return acc + road.road.price * road.quantity * 100;
+            return acc + road.roadsVeichles.price * road.quantity * 100;
+          }, 0);
+          const totalMeal = this.meals.reduce((acc: any, meal: any) => {
+            return acc + meal.quantity * meal.price * 100;
           }, 0);
 
           this.total =
-            totalAccomodation / 100 + totalActivity / 100 + totalRoads / 100;
+            totalAccomodation / 100 +
+            totalActivity / 100 +
+            totalRoads / 100 +
+            totalMeal / 100;
 
-          this.roads = this.cart?.roads.reduce((group: any, road: any) => {
-            const index = group.findIndex((g: any) =>
-              g.data.some(
-                (el: any) => !compareDates(el.startDate, road.startDate)
-              )
-            );
-            if (index > -1) {
-              group[index].data.push(road);
-            } else {
-              group.push({ date: road.startDate, data: [road] });
-            }
-            return group;
-          }, []);
+          console.log('this.total', this.total);
         })
       )
       .subscribe();
@@ -137,18 +168,23 @@ export class PersonalCartViewComponent {
   }
 
   update() {
-    this.router.navigate(['personal', 'carts', 'edit', this.cart.id]);
+    this.fromBackoffice
+      ? this.router.navigate(['cart', 'edit', this.cart.id])
+      : this.router.navigate(['personal', 'carts', 'edit', this.cart.id]);
   }
 
   complete() {
     this.confirmationService.confirm({
       message:
-        'Una volta confermata la trasferta non potrà più essere modificata e verrà inviata una mail al nostro staff. Si intende procedere?',
+        'Una volta confermata la trasferta verrà inviata una mail al nostro staff. Si intende procedere?',
       header: 'Conferma',
       icon: 'pi pi-exclamation-triangle',
       accept: () => {
         this.cartApiService
-          .update(this.cart.id, { status: EStatusCart.PENDING })
+          .update(this.cart.id, {
+            status: EStatusCart.PENDING,
+            onlyStatus: true,
+          })
           .pipe(
             take(1),
             tap(() => {
@@ -165,8 +201,48 @@ export class PersonalCartViewComponent {
     });
   }
 
+  updateStatus(status: EStatusCart) {
+    const messages = {
+      [EStatusCart.CANCELLED]:
+        'Sei sicuro di voler annullare questa trasferta?',
+      [EStatusCart.CONFIRMED]:
+        'Sei sicuro di voler confermare questa trasferta?',
+      [EStatusCart.DEPOSIT]:
+        "Sei sicuro di voler confermare il pagamento dell'acconto?",
+      [EStatusCart.COMPLETED]:
+        'Sei sicuro di voler completare questa trasferta?',
+      [EStatusCart.DRAFT]: '',
+      [EStatusCart.PENDING]: '',
+    };
+    this.confirmationService.confirm({
+      message: messages[status],
+      header: 'Conferma',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.cartApiService
+          .update(this.cartId, { status, onlyStatus: true })
+          .pipe(
+            take(1),
+            tap(() => {
+              this.loadCart();
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Operazione completata',
+              });
+            })
+          )
+          .subscribe();
+      },
+    });
+  }
+
   download() {
     this.isDownloading = true;
+    this.messageService.add({
+      severity: 'success',
+      summary: 'Download iniziato',
+      detail: ' Il file verrà scaricato nel browser tra qualche secondo!',
+    });
     this.cartApiService
       .downloadPdf(this.cartId)
       .pipe(
@@ -180,8 +256,12 @@ export class PersonalCartViewComponent {
           link.click();
         })
       )
-      .subscribe((data) => {
-        //this.blob = new Blob([data], {type: 'application/pdf'});
-      });
+      .subscribe();
+  }
+
+  backToCarts() {
+    this.fromBackoffice
+      ? this.router.navigate(['cart'])
+      : this.router.navigate(['personal', 'carts']);
   }
 }
